thanh toan

+++ app/Http/Controllers/Admin/BDSController.php
@@ -127,14 +127,13 @@ class BDSController extends Controller
             'ghichu',
             'toado'         => 'required',
         ], $this->model->messages);
-
-        $this->model->find($id)->update($request->only('id_csh','id_loaibds','id_tp','ten_bds','alias','diachi_bds','soCNQSDD','ghichu','toado','hinhthuc','gia'));
-
-
-        $bds = $this->model->find($id);
         
-        if($request->hinhanh != null && $bds->hinhanh != null){
-            if(file_exists("upload/bds/".$bds->hinhanh))
+        $bds = $this->model->find($id);
+
+        $bds->update($request->only('id_csh','id_loaibds','id_tp','ten_bds','alias','diachi_bds','soCNQSDD','ghichu','toado','hinhthuc','gia'));
+        
+        if(!empty($request->hinhanh)){
+            if(!empty($bds->hinhanh) && file_exists("upload/bds/".$bds->hinhanh))
             {
                 unlink("upload/bds/".$bds->hinhanh);
             }
@@ -142,15 +141,9 @@ class BDSController extends Controller
             $bds->hinhanh = $hinhanh;
             $bds->update();
             $request->file('hinhanh')->move('upload/bds',$hinhanh);
-        // }else{
-        //     $hinhanh = $request->alias.'.'.$request->file('hinhanh')->getClientOriginalExtension();
-        //     $bds->hinhanh = $hinhanh;
-        //     $bds->update();
-        //     $request->file('hinhanh')->move('upload/bds',$hinhanh);
-        }   
+        }
 
         session()->flash('flash_message', 'Cập nhật dữ liệu thành công');
-        // return redirect()->back();
         return redirect()->route('bds.index');
     }
